+++ daemail/__main__.py
@@ -50,7 +50,7 @@ def main():
                         help='Use UTC timestamps')
 
     sendarg = parser.add_mutually_exclusive_group()
-    sendarg.add_argument('-m', '--mail-cmd', metavar='COMMAND',
+    sendarg.add_argument('-s', '--sendmail', metavar='COMMAND',
                         help='Command for sending e-mail')
     sendarg.add_argument('--mbox', help='Append e-mail to this mbox file')
     sendarg.add_argument('--smtp-host', metavar='HOST',
@@ -104,7 +104,7 @@ def main():
     elif args.mbox is not None:
         sender = senders.MboxSender(os.path.join(pwd, args.mbox))
     else:
-        sender = senders.CommandSender(args.mail_cmd)
+        sender = senders.CommandSender(args.sendmail)
 
     mailer = CommandMailer(
         encoding=args.encoding,